Fixed Loading and implemented tests for repository class

+++ app/src/main/java/com/kwdcurrencyconverter/data/models/CurrencySymbolsResponse.kt
@@ -8,6 +8,6 @@ data class CurrencySymbolsResponse (
     @SerializedName("rates")
     val extras: Map<String, Double>,
     @SerializedName("timestamp")
-    val timestamp: Double
+    val timestamp: Long
 
 )
